made slider more general
