+++ Assets/HotmaxEditorVR/presentation/common/slider.mat
