+++ Assets/HotmaxEditorVR/presentation/common/sliderSystem.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 4a1619128aede10468a5011f8d172081
-timeCreated: 1518722437
+guid: d062e5b55cb6d3c46b00de30d85128bd
+timeCreated: 1521493248
 licenseType: Pro
 MonoImporter:
   externalObjects: {}
