+++ Assets/HotmaxEditorVR/presentation/common/sliderSystem.prefab
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:0f6c37b47f3e81649544912fd46cba3aa0af7640911ea25fb3810b9b1e7f1d92
+size 14536
