+++ Assets/HotmaxEditorVR/presentation/offsetCalibration/offsetCalibration.unity
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:45e37cefb0c109464faace2a8d3fab059ab7d1333f64fdfed69c297a1d55ac45
-size 670463
+oid sha256:875515f9c909533cc380de525870d0e33f89d13830689ea343b611bf68e00c53
+size 660428
